+++ sdk/storage/azfile/directory/models.go
@@ -10,7 +10,6 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azfile/file"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azfile/internal/exported"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azfile/internal/generated"
-	"github.com/Azure/azure-sdk-for-go/sdk/storage/azfile/internal/shared"
 	"reflect"
 )
 
@@ -22,6 +21,9 @@ type SharedKeyCredential = exported.SharedKeyCredential
 func NewSharedKeyCredential(accountName, accountKey string) (*SharedKeyCredential, error) {
 	return exported.NewSharedKeyCredential(accountName, accountKey)
 }
+
+// DestinationLeaseAccessConditions contains optional parameters to access the destination directory.
+type DestinationLeaseAccessConditions = generated.DestinationLeaseAccessConditions
 
 // ---------------------------------------------------------------------------------------------------------------------
 
@@ -35,16 +37,17 @@ type CreateOptions struct {
 	Metadata map[string]*string
 }
 
-func (o *CreateOptions) format() (string, *generated.DirectoryClientCreateOptions) {
+func (o *CreateOptions) format() *generated.DirectoryClientCreateOptions {
 	if o == nil {
-		return shared.FileAttributesDirectory, nil
+		return nil
 	}
 
-	fileAttributes, fileCreationTime, fileLastWriteTime, fileChangeTime := o.FileSMBProperties.Format(true, shared.FileAttributesDirectory)
+	fileAttributes, fileCreationTime, fileLastWriteTime, fileChangeTime := o.FileSMBProperties.Format(true)
 
 	permission, permissionKey := o.FilePermissions.Format()
 
 	createOptions := &generated.DirectoryClientCreateOptions{
+		FileAttributes:    fileAttributes,
 		FileChangeTime:    fileChangeTime,
 		FileCreationTime:  fileCreationTime,
 		FileLastWriteTime: fileLastWriteTime,
@@ -53,7 +56,7 @@ func (o *CreateOptions) format() (string, *generated.DirectoryClientCreateOption
 		Metadata:          o.Metadata,
 	}
 
-	return fileAttributes, createOptions
+	return createOptions
 }
 
 // ---------------------------------------------------------------------------------------------------------------------
@@ -65,6 +68,55 @@ type DeleteOptions struct {
 
 func (o *DeleteOptions) format() *generated.DirectoryClientDeleteOptions {
 	return nil
+}
+
+// ---------------------------------------------------------------------------------------------------------------------
+
+// RenameOptions contains the optional parameters for the Client.Rename method.
+type RenameOptions struct {
+	// FileSMBProperties contains the optional parameters regarding the SMB/NTFS properties for a file.
+	FileSMBProperties *file.SMBProperties
+	// FilePermissions contains the optional parameters for the permissions on the file.
+	FilePermissions *file.Permissions
+	// IgnoreReadOnly specifies whether the ReadOnly attribute on a pre-existing destination file should be respected.
+	// If true, rename will succeed, otherwise, a previous file at the destination with the ReadOnly attribute set will cause rename to fail.
+	IgnoreReadOnly *bool
+	// A name-value pair to associate with a file storage object.
+	Metadata map[string]*string
+	// ReplaceIfExists specifies that if the destination file already exists, whether this request will overwrite the file or not.
+	// If true, rename will succeed and will overwrite the destination file. If not provided or if false and the destination file does exist,
+	// the request will not overwrite the destination file.
+	// If provided and the destination file does not exist, rename will succeed.
+	ReplaceIfExists *bool
+	// DestinationLeaseAccessConditions contains optional parameters to access the destination directory.
+	DestinationLeaseAccessConditions *DestinationLeaseAccessConditions
+}
+
+func (o *RenameOptions) format() (*generated.DirectoryClientRenameOptions, *generated.DestinationLeaseAccessConditions, *generated.CopyFileSMBInfo) {
+	if o == nil {
+		return nil, nil, nil
+	}
+
+	fileAttributes, fileCreationTime, fileLastWriteTime, fileChangeTime := o.FileSMBProperties.Format(true)
+
+	permission, permissionKey := o.FilePermissions.Format()
+
+	renameOpts := &generated.DirectoryClientRenameOptions{
+		FilePermission:    permission,
+		FilePermissionKey: permissionKey,
+		IgnoreReadOnly:    o.IgnoreReadOnly,
+		Metadata:          o.Metadata,
+		ReplaceIfExists:   o.ReplaceIfExists,
+	}
+
+	smbInfo := &generated.CopyFileSMBInfo{
+		FileAttributes:    fileAttributes,
+		FileChangeTime:    fileChangeTime,
+		FileCreationTime:  fileCreationTime,
+		FileLastWriteTime: fileLastWriteTime,
+	}
+
+	return renameOpts, o.DestinationLeaseAccessConditions, smbInfo
 }
 
 // ---------------------------------------------------------------------------------------------------------------------
@@ -95,23 +147,24 @@ type SetPropertiesOptions struct {
 	FilePermissions *file.Permissions
 }
 
-func (o *SetPropertiesOptions) format() (string, *generated.DirectoryClientSetPropertiesOptions) {
+func (o *SetPropertiesOptions) format() *generated.DirectoryClientSetPropertiesOptions {
 	if o == nil {
-		return shared.DefaultPreserveString, nil
+		return nil
 	}
 
-	fileAttributes, fileCreationTime, fileLastWriteTime, fileChangeTime := o.FileSMBProperties.Format(true, shared.DefaultPreserveString)
+	fileAttributes, fileCreationTime, fileLastWriteTime, fileChangeTime := o.FileSMBProperties.Format(true)
 
 	permission, permissionKey := o.FilePermissions.Format()
 
 	setPropertiesOptions := &generated.DirectoryClientSetPropertiesOptions{
+		FileAttributes:    fileAttributes,
 		FileChangeTime:    fileChangeTime,
 		FileCreationTime:  fileCreationTime,
 		FileLastWriteTime: fileLastWriteTime,
 		FilePermission:    permission,
 		FilePermissionKey: permissionKey,
 	}
-	return fileAttributes, setPropertiesOptions
+	return setPropertiesOptions
 }
 
 // ---------------------------------------------------------------------------------------------------------------------