+++ sdk/storage/azfile/directory/responses.go
@@ -14,6 +14,12 @@ type CreateResponse = generated.DirectoryClientCreateResponse
 // DeleteResponse contains the response from method Client.Delete.
 type DeleteResponse = generated.DirectoryClientDeleteResponse
 
+// RenameResponse contains the response from method Client.Rename.
+type RenameResponse struct {
+	generated.DirectoryClientRenameResponse
+	Client *Client
+}
+
 // GetPropertiesResponse contains the response from method Client.GetProperties.
 type GetPropertiesResponse = generated.DirectoryClientGetPropertiesResponse