+++ sdk/storage/azfile/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Updated service version to STG 87(2022-11-02).
 * Added OAuth support.
+* Added [Rename Directory API](https://learn.microsoft.com/rest/api/storageservices/rename-directory).
+* Added [Rename File API](https://learn.microsoft.com/rest/api/storageservices/rename-file).
 
 ### Breaking Changes